Servers/requirements.txt
async-timeout==4.0.2
bidict==0.22.1
blinker==1.6.2
cassandra-driver==3.28.0
click==8.1.4
colorama==0.4.6
confluent-kafka==2.2.0
Flask==2.3.2
flask-kafka==0.0.3
Flask-SocketIO==5.3.4
geomet==0.2.1.post1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
python-engineio==4.5.1
python-socketio==5.8.0
redis==4.6.0
simple-websocket==0.10.1
six==1.16.0
Werkzeug==2.3.6
wsproto==1.2.0
